+++ .github/ci.sh
@@ -99,7 +99,12 @@ install_solvers() {
   chmod +x $BIN/*
   export PATH=$BIN:$PATH
   echo "$BIN" >> "$GITHUB_PATH"
-  is_exe "$BIN" z3 && is_exe "$BIN" cvc4 && is_exe "$BIN" yices
+  is_exe   "$BIN" z3 && \
+    is_exe "$BIN" cvc4 && \
+    is_exe "$BIN" cvc5 && \
+    is_exe "$BIN" boolector && \
+    is_exe "$BIN" bitwuzla && \
+    is_exe "$BIN" yices
 }
 
 install_system_deps() {